+++ src/orchestrator/deploy/amazon.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"lorhammer/src/tools"
+	"os/exec"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
@@ -26,9 +27,21 @@ type amazonImpl struct {
 	DistantConfig    json.RawMessage `json:"distantConfig"`
 
 	ec2Client       *ec2.EC2
-	distantDeployer *distantImpl
+	distantDeployer *amazonDistantImpl
 	mqttAddress     string
 	instancesID     []*string
+}
+
+type amazonDistantImpl struct {
+	SSHKeyPath        string `json:"sshKeyPath"`
+	User              string `json:"user"`
+	IPServer          string `json:"ipServer"`
+	PathFile          string `json:"pathFile"`
+	PathWhereScp      string `json:"pathWhereScp"`
+	BeforeCmd         string `json:"beforeCmd"`
+	AfterCmd          string `json:"afterCmd"`
+	NbDistantToLaunch int    `json:"nbDistantToLaunch"`
+	cmdFabric         func(name string, arg ...string) *exec.Cmd
 }
 
 func newAmazonFromJSON(serialized json.RawMessage, mqttClient tools.Mqtt) (deployer, error) {
@@ -44,11 +57,72 @@ func newAmazonFromJSON(serialized json.RawMessage, mqttClient tools.Mqtt) (deplo
 	}
 
 	client.ec2Client = ec2.New(s)
-	if client.distantDeployer, err = newDistantImpl(client.DistantConfig); err != nil {
+	if client.distantDeployer, err = newAmazonDistantImpl(client.DistantConfig); err != nil {
 		return nil, err
 	}
 	client.mqttAddress = mqttClient.GetAddress()
 	return client, nil
+}
+
+func newAmazonDistantImpl(serialized json.RawMessage) (*amazonDistantImpl, error) {
+	d := &amazonDistantImpl{}
+	if err := json.Unmarshal(serialized, d); err != nil {
+		return nil, err
+	}
+	d.cmdFabric = exec.Command
+	return d, nil
+}
+
+func (distant *amazonDistantImpl) Deploy() error {
+
+	c := fmt.Sprintf("%s@%s:%s", distant.User, distant.IPServer, distant.PathWhereScp)
+	logDistant.WithField("cmd", "scp -q -i "+distant.SSHKeyPath+" -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null "+distant.PathFile+" "+c).Info("Will exec cmd")
+	cmd := distant.cmdFabric("scp", "-q", "-i", distant.SSHKeyPath, "-o", "StrictHostKeyChecking=no", "-o", "UserKnownHostsFile=/dev/null", distant.PathFile, c)
+	if stdoutStderr, err := cmd.CombinedOutput(); err != nil {
+		logDistant.WithField("output", fmt.Sprintf("%s", stdoutStderr)).Info("Scp output")
+		return err
+	}
+	return nil
+}
+
+func (distant *amazonDistantImpl) runCmd(cmd string) error {
+	errs := distantRunError{Errors: make([]error, 0)}
+	ip := fmt.Sprintf("%s@%s", distant.User, distant.IPServer)
+	logDistant.WithField("cmd", "ssh -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null "+ip+" "+cmd).Info("Will exec cmd")
+
+	chanErr := make(chan error)
+	defer close(chanErr)
+
+	for i := 0; i < distant.NbDistantToLaunch; i++ {
+		go func() {
+			cmd := distant.cmdFabric("ssh", "-q", "-o", "StrictHostKeyChecking=no", "-o", "UserKnownHostsFile=/dev/null", ip, cmd)
+			if stdoutStderr, err := cmd.CombinedOutput(); err != nil {
+				logDistant.WithField("output", fmt.Sprintf("%s", stdoutStderr)).Info("Ssh output")
+				chanErr <- err
+			} else {
+				chanErr <- nil
+			}
+		}()
+	}
+
+	for i := 0; i < distant.NbDistantToLaunch; i++ {
+		if err := <-chanErr; err != nil {
+			errs.Errors = append(errs.Errors, err)
+		}
+	}
+
+	if len(errs.Errors) > 0 {
+		return errs
+	}
+	return nil
+}
+
+func (distant *amazonDistantImpl) RunAfter() error {
+	err := distant.runCmd(distant.AfterCmd)
+	if err != nil {
+		return err
+	}
+	return nil
 }
 
 func (client *amazonImpl) RunBefore() error {